+++ server/src/main/java/org/apache/iotdb/db/utils/datastructure/DoubleTVList.java
@@ -94,11 +94,13 @@ public class DoubleTVList extends TVList {
 
   @Override
   public void sort() {
-    if (sortedTimestamps == null || sortedTimestamps.length < size) {
+    if (sortedTimestamps == null
+        || sortedTimestamps.length < PrimitiveArrayManager.getArrayRowCount(size)) {
       sortedTimestamps =
           (long[][]) PrimitiveArrayManager.createDataListsByType(TSDataType.INT64, size);
     }
-    if (sortedValues == null || sortedValues.length < size) {
+    if (sortedValues == null
+        || sortedValues.length < PrimitiveArrayManager.getArrayRowCount(size)) {
       sortedValues =
           (double[][]) PrimitiveArrayManager.createDataListsByType(TSDataType.DOUBLE, size);
     }